% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiple_lvs.R
\name{.run_multiple_lvs}
\alias{.run_multiple_lvs}
\title{Run sRDA with n > 1 multiple latnet variables}
\usage{
.run_multiple_lvs(
  X,
  Y,
  penalization,
  lambdas,
  nonzeros,
  n_lvs,
  tolerance,
  max_iteration,
  cv_n_folds,
  parallel,
  seed
)
}
\arguments{
\item{X}{explanatory matrix or data frame, n x p}

\item{Y}{Response matrix, n x q}

\item{penalization}{penalization methods}

\item{lambdas}{x}

\item{nonzeros}{x}

\item{n_lvs}{x}

\item{tolerance}{x}

\item{max_iteration}{x}

\item{cv_n_folds}{x}

\item{parallel}{x}

\item{seed}{x}
}
\value{
x
}
\description{
This is an internal function.
}
\examples{
x
}
