% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stability_test.R
\name{run_stability_test}
\alias{run_stability_test}
\title{Bootstrapping test for the significance of features in sRDA}
\usage{
run_stability_test(rda, method, n_bootstrap = 1000, ci = 0.95)
}
\arguments{
\item{rda}{x}

\item{method}{x}

\item{n_bootstrap}{Number of bootstrapping}

\item{ci}{Confidence interval}
}
\value{
a result data frame
}
\description{
Bootstrapping test for the significance of features in sRDA
}
\examples{
x

}
