% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srda.R
\name{srda}
\alias{srda}
\title{Sparse Redundancy Analysis (sRDA)}
\usage{
srda(
  explanatory,
  response,
  penalization,
  lambdas = 1,
  nonzeros = 1,
  n_lvs = 1,
  max_iteration = 100,
  tolerance = 1e-20,
  parallel = FALSE,
  cv_n_folds,
  seed
)
}
\arguments{
\item{explanatory}{explanatory matrix or data frame, n x p}

\item{response}{response matrix or data frame, n x q}

\item{penalization}{The penalization method: "enet", "ust", "none"}

\item{lambdas}{The ridge penalty parameter of the explanatory latent variable (Xi) used for enet}

\item{nonzeros}{The number of non-zero weights of the explanatory latent variable (Xi) used for enet or ust}

\item{n_lvs}{Number of latent variables}

\item{max_iteration}{Maximum number of iterations}

\item{tolerance}{Convergence criteria}

\item{parallel}{logical; run cross-validation in parallel or not}

\item{cv_n_folds}{x}

\item{seed}{seed number, default is 1234}
}
\value{
A sRDA object
}
\description{
This package is a rewrite of the sRDA package available on CRAN. Please visit
the original package on https://cran.r-project.org/package=sRDA and their
paper for more details. If you use this package, please consider cite the
original paper of sRDA.
}
\examples{
x
}
